Initial implememtation of functions to propagate and type check variables
:Release Notes:
Initial implememtation of functions to propagate and type check variables

:Detailed Notes:
Initial implememtation of functions to propagate and type check variables

A SCRATCH area in an object holds the values propagated up to it.
The type checking is done on fields in the SCRATCH area, only if they hold multiple values.

:Testing Performed:
The output can be found in analysis.txt

:QA Notes:

:Issues Addressed: